client side binding perf
- on the client we do not use a linked list like we do on the
server or storing the list of binding listeners for a given
property.   The 'splice' call in the array was showing up
in the stack traces a lot so eliminated that by storing a null
on delete and code to find the empty slot on insert.

- bug fix for SelectorBinding revalidate not listening on
property change (and getting an error on remove listeners)

- anyChanges return flag for refreshList in Repeatcomponent

- getName method for js_Element_c so it can act like a normal java
class