added EditablePage class for plugging in management UI
Before any <html> page would by default pick up the "editability"
but this makes it awkward to do JSP pages like stateViewer and
things which need to run in request scope, stateless mode. We could
device a way that they extend a new page but it seems better
overall to make manageability an 'opt in' feature, especially
since you need to define the right css classes to get the display
to work right.