comments and diagnostic code
- helped find a bug where a scheduled job was running on the
server and modifying a page-type with a 'refresh' scheduled
to run on the next request.  Ultimately we need to get rid
of that "run on the next request" hook and it does not work
with sync - only the next page request so it meant we would
never refresh the server tags once we hit this problem.