log4j prod layer
removed jetty v8 compat layer because it was messing up
the types in the IDE

use produceJar = true for jetty by default to make it easier to
deploy simple web apps as a single jar file.  Ideally I'd also
like to offer a way to just copy the libraries and do an rsync so
that only changed files are transferred, but this is another great
way to do it so that there's only a single jar file to roll in/out for
the entire web app, plus dependencies.