fixed problem with java 13 and relative path to the web directory
In java 8, the System.setProperty("user.dir", ..) call would
change the behavior for a new 'getCanicalPath()' call and in java 13
it no longer does. It only affects the case where we change to the buildDir
after the build and run the 'main' as part of the dynamic runtime but it's
easy enough to workaround by just using the system property to form the
absolute path.