no longer changing Object.prototype
since it breaks 3rd party JS code that does for (x in y) without
doing a hasOnlyProperty such as jquery and codemirror.  The fix for
codemirror was just one line so we patched it for a while but jquery
had a similar problem and it seemed worth while.

Perhaps the better fix for this would have been for hashCode/equals and
getClass method calls to be code-generated as static calls passing in
the object as an argument - e.g. jv_Object_c.objHashCode(obj) but for now
we are changing the prototype on all of the subclasses and ensure that
any Java types we use extend jv_Object.