reimplement resize event to use ResizeObserver
this new browser feature is a better alternative so we get
notified when the size properties might have changed. If this doesn't
exist, we fallback to checking for size changes on the window resize
but this won't account for refreshes within the page that change the
size of an interior element

another approach we could use here if resize observer is not perfect
is to add a check size call after each UI interaction - we just see
if the size has changed since the last time and send events for
any size properties that have changed so it's not too expensive.