moved index.schtml to devIndex.schtml
only use devIndex.schtml when in devMode and only if
there is not another page already registered for the "/"
page.

this is to reduce the chance the dev index makes it to
production. Although we can override the index, it's kind of
awkward since you have to use tagMerge="replace" and in general
the complexity of having to override one class with another is
something I like to avoid except in extreme cases where you really
need to do it.