turning off sync of this property by default
since in the ec bundle we bind it to a synchronized property and it's not
a good idea to sync the same state in multiple properties. The bug was
probably different though so this might be ok. Ultimately we need to make
sure that we don't set the 'content' property to a stale value, so changes
should not come back from the server to a property that's directly bound
to the input of the text field. In the ec example, we'd sync longDescHtml
that needed to be filtered, then copied into category.longDesc but that change
got sync'd back and updated the text editor - which might have subsequently
been changed again so the text would not match.

Now the longDescHtml property is used for the forward binding and it's only
set when the new product is selected